tests: Split libatasmart and smartmontool tests
